gui_main_flask/requirements.txt
bidict==0.21.2
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-SocketIO==5.0.1
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
networkx==2.5
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==4.0.1
python-socketio==5.1.0
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.1
six==1.15.0
tifffile==2021.3.5
Werkzeug==1.0.1
